separate warning decorators
